--- app/src/main/java/com/practicum/playlistmaker/data/model/NetworkResponse.kt
@@ -1,5 +1,0 @@
-package com.practicum.playlistmaker.data.model
-
-open class NetworkResponse{
-    var resultCode = 0
-}
