--- app/src/main/java/com/practicum/playlistmaker/data/model/ResponseTracks.kt
@@ -1,5 +1,0 @@
-package com.practicum.playlistmaker.data.model
-
-import com.google.gson.annotations.SerializedName
-
-data class ResponseTracks(@SerializedName("results") val foundTracks: List<TrackDTO>): NetworkResponse()
