--- app/src/main/java/com/practicum/playlistmaker/data/network/ITunesNetworkClient.kt
@@ -1,7 +1,0 @@
-package com.practicum.playlistmaker.data.network
-
-import com.practicum.playlistmaker.data.model.NetworkResponse
-
-interface ITunesNetworkClient {
-    fun getTracks(currency: String): NetworkResponse
-}
