--- app/src/main/java/com/practicum/playlistmaker/domain/api/InteractorHistory.kt
@@ -1,9 +1,0 @@
-package com.practicum.playlistmaker.domain.api
-
-import com.practicum.playlistmaker.domain.model.Track
-
-interface InteractorHistory {
-    fun getHistory(): MutableList<Track>
-    fun add(item: Track)
-    fun clear()
-}
