--- app/src/main/java/com/practicum/playlistmaker/domain/model/Resource.kt
@@ -1,6 +1,0 @@
-package com.practicum.playlistmaker.domain.model
-
-sealed interface Resource<T> {
-    data class Success<T>(val data: T): Resource<T>
-    data class Error<T>(val msg: String): Resource<T>
-}
