+++ app/src/main/java/com/practicum/playlistmaker/search/data/model/NetworkResponse.kt
@@ -1,0 +1,5 @@
+package com.practicum.playlistmaker.search.data.model
+
+open class NetworkResponse{
+    var resultCode = 0
+}
