+++ app/src/main/java/com/practicum/playlistmaker/search/data/model/ResponseTracks.kt
@@ -1,0 +1,5 @@
+package com.practicum.playlistmaker.search.data.model
+
+import com.google.gson.annotations.SerializedName
+
+data class ResponseTracks(@SerializedName("results") val foundTracks: List<TrackDTO>): NetworkResponse()
