+++ app/src/main/java/com/practicum/playlistmaker/search/data/model/TrackDTO.kt
@@ -1,0 +1,13 @@
+package com.practicum.playlistmaker.search.data.model
+
+data class TrackDTO(
+    val trackName: String, // Название композиции
+    val artistName: String, // Имя исполнителя
+    val trackTimeMillis: String, // Продолжительность трека
+    val artworkUrl100: String, // Ссылка на изображение обложки
+    val collectionName: String, // Альбом
+    val releaseDate: String, // Год релиза
+    val primaryGenreName: String, // Жанр
+    val country: String, // Страна исполнителя
+    val previewUrl: String //Ссылка на превью(первые 30 сек) трека
+)
