+++ app/src/main/java/com/practicum/playlistmaker/search/data/network/ITunesNetworkClient.kt
@@ -1,0 +1,7 @@
+package com.practicum.playlistmaker.search.data.network
+
+import com.practicum.playlistmaker.search.data.model.NetworkResponse
+
+interface ITunesNetworkClient {
+    fun getTracks(currency: String): NetworkResponse
+}
