+++ app/src/main/java/com/practicum/playlistmaker/search/domain/api/RepositoryHistory.kt
@@ -1,0 +1,9 @@
+package com.practicum.playlistmaker.search.domain.api
+
+import com.practicum.playlistmaker.search.domain.model.Track
+
+interface RepositoryHistory {
+    fun getHistory(): MutableList<Track>
+    fun add(item: Track)
+    fun clear()
+}
