+++ app/src/main/java/com/practicum/playlistmaker/search/domain/api/RepositoryNetwork.kt
@@ -1,0 +1,8 @@
+package com.practicum.playlistmaker.search.domain.api
+
+import com.practicum.playlistmaker.search.domain.model.Resource
+import com.practicum.playlistmaker.search.domain.model.Track
+
+interface RepositoryNetwork {
+    fun getTracks(currency: String): Resource<List<Track>>
+}
