+++ app/src/main/java/com/practicum/playlistmaker/search/domain/consumer/Consumer.kt
@@ -1,0 +1,5 @@
+package com.practicum.playlistmaker.search.domain.consumer
+
+interface Consumer<T> {
+    fun consume(data: ConsumerData<T>)
+}
