+++ app/src/main/java/com/practicum/playlistmaker/search/domain/consumer/ConsumerData.kt
@@ -1,0 +1,6 @@
+package com.practicum.playlistmaker.search.domain.consumer
+
+sealed interface ConsumerData<T>{
+    data class Data<T>(val value: T): ConsumerData<T>
+    data class Error<T>(val msg: String): ConsumerData<T>
+}
