+++ app/src/main/java/com/practicum/playlistmaker/search/domain/model/Resource.kt
@@ -1,0 +1,6 @@
+package com.practicum.playlistmaker.search.domain.model
+
+sealed interface Resource<T> {
+    data class Success<T>(val data: T): Resource<T>
+    data class Error<T>(val msg: String): Resource<T>
+}
