--- app/src/main/java/com/practicum/playlistmaker/ui/MainActivity.kt
@@ -1,49 +1,0 @@
-package com.practicum.playlistmaker.ui
-
-import android.content.Intent
-import android.os.Bundle
-import android.widget.Button
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import com.practicum.playlistmaker.R
-import com.practicum.playlistmaker.ui.search.SearchActivity
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        enableEdgeToEdge()
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
-            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            val viewPadding = resources.getDimensionPixelSize(R.dimen.padding)
-            v.setPadding(
-                systemBars.left + viewPadding,
-                systemBars.top + viewPadding,
-                systemBars.right + viewPadding,
-                systemBars.bottom + viewPadding)
-            insets
-        }
-
-        val button_search = findViewById<Button>(R.id.button_search)
-        val button_library = findViewById<Button>(R.id.button_library)
-        val button_settings = findViewById<Button>(R.id.button_settings)
-
-        button_search.setOnClickListener {
-            val searchIntent = Intent(this@MainActivity, SearchActivity::class.java)
-            startActivity(searchIntent)
-        }
-
-        button_library.setOnClickListener {
-            val libraryIntent = Intent(this@MainActivity, LibraryActivity::class.java)
-            startActivity(libraryIntent)
-        }
-
-        button_settings.setOnClickListener {
-            val settingsIntent = Intent(this@MainActivity, SettingsActivity::class.java)
-            startActivity(settingsIntent)
-        }
-
-    }
-}
