+++ app/src/main/res/layout/activity_library.xml
@@ -5,6 +5,6 @@
     android:id="@+id/main_library"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.LibraryActivity">
+    tools:context=".LibraryActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
