+++ app/src/main/res/layout/activity_main.xml
@@ -9,7 +9,7 @@
     android:orientation="vertical"
     android:padding="@dimen/padding"
     android:id="@+id/main"
-    tools:context=".ui.MainActivity">
+    tools:context=".main.ui.MainActivity">
 
 
     <TextView
